+++ packages/plugin-packages/mashroom-security-provider-ldap/src/mashroom-bootstrap.ts
@@ -8,7 +8,8 @@ import type {MashroomSecurityProviderPluginBootstrapFunction} from '@mashroom/ma
 const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
     const {
         loginPage, userSearchFilter, groupSearchFilter, extraDataMapping, secretsMapping,
-        groupToRoleMapping: groupToRoleMappingPath, serverUrl, ldapConnectTimeout, ldapTimeout,
+        groupToRoleMapping: groupToRoleMappingPath, userToRoleMapping: userToRoleMappingPath,
+        serverUrl, ldapConnectTimeout, ldapTimeout,
         tlsOptions, baseDN, bindDN, bindCredentials, authenticationTimeoutSec,
     } = pluginConfig;
 
@@ -23,7 +24,7 @@ const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (plugin
     });
 
     return new MashroomLdapSecurityProvider(loginPage, userSearchFilter, groupSearchFilter, extraDataMapping, secretsMapping,
-        groupToRoleMappingPath, ldapClient, serverRootFolder, authenticationTimeoutSec, loggerFactory);
+        groupToRoleMappingPath, userToRoleMappingPath, ldapClient, serverRootFolder, authenticationTimeoutSec, loggerFactory);
 };
 
 export default bootstrap;