+++ CHANGELOG.md
@@ -11,6 +11,10 @@
       * Instead of returning *result.reject* you can now call res.sendStatus() yourself and just return *result.responseHandled* as
         a hint that the proxy doesn't have to do anything.
  * HTTP Proxy: Fixed a problem with special characters in target URIs
+ * LDAP Security Provider: Added an optional user to role mapping which is useful if you want to give just a
+   specific user access to a Portal page and such.
+ * Security: The *MashroomSecurityService.login()* method also returns now a reason if possible (e.g. Invalid credentials).
+   This works at the moment for the Simple Provider and the LDAP Provider with Active Directory.
  * Portal: Fixed the client log handler (didn't terminate correctly)
 
 ## 1.6.3 (December 20, 2020)